+++ yalafi/packages/amsthm.py
@@ -3,7 +3,7 @@
 #   YaLafi module for LaTeX package amsthm
 #
 
-from yalafi.defs import ModParm, Environ
+from yalafi.defs import ModParm, Environ, SpaceToken, TextToken
 
 require_packages = []
 
@@ -20,12 +20,19 @@ def modify_parameters(parms):
 
     environments = [
 
-        Environ(parms, 'proof', args='O',
-                            # Parser.expand_arguments() may skip space
-                            repl='#1.\n', defaults=[parms.proof_name]),
+        Environ(parms, 'proof', args='O', repl=h_proof),
 
     ]
 
     return ModParm(macros_latex=macros_latex, macros_python=macros_python,
                         environments=environments)
+
+def h_proof(parser, buf, mac, args, pos):
+    if args[0]:
+        ret = args[0]
+    else:
+        ret = [TextToken(pos, parser.parms.lang_context().proof_name,
+                        pos_fix=True)]
+    return ret + [TextToken(ret[-1].pos, '.', pos_fix=True),
+                        SpaceToken(ret[-1].pos, '\n', pos_fix=True)]